Error stated in #58 is no longer showing
Perhaps it's a naive or not-so-nuanced fix, but removing the line about which Neovim complains has "fixed" the issue for me. This is my first time trying to contribute to a real-world project. Further, I don't do much with Lua outside of my own configuration. Any guidance, criticisms, or clarifications will be happily accepted and implemented, if need be. Thanks!